Setup to build on Netlify
This sets up the Netlify configuration for this spec site, and enables us to
migrate from GitHub pages to the standard way we deploy website properties
across Apollo: Netlify.

The site has been added in Netlify already at:
  https://app.netlify.com/sites/apollo-specs-join/

This also changes the CSS and logo-rendering paths to reference the specs
site using its FQDN as a full URL to enable the site to work anywhere it is
mounted, including on localhost and on Netlify previews.

In the future, perhaps we can have this bundled into the spec site itself
so the site can stand on its own in the same way that our docs do, and not
require external artifacts (and have our deploys be more hermetically
sealed / permadurable on Netlify).